fix: add base interaction or add and edit interactions
