+++ app/src/main/java/com/london/tudee/presentation/base/HomeInteractions.kt
@@ -4,19 +4,19 @@ import com.london.tudee.domain.entities.Category
 import com.london.tudee.domain.entities.Priority
 import com.london.tudee.domain.entities.Task
 
-interface HomeInteractions {
+interface HomeInteractions:AddOrEditInteractions {
     fun onThemeSwitched(isDarkMode: Boolean)
     fun loadCategories()
-    fun updateTitle(title: String)
-    fun updateDescription(description: String)
-    fun updateSelectedDate(date: Long)
-    fun updateSelectedPriority(priority: Priority)
-    fun updateSelectedCategory(category: Category)
-    fun showDatePicker()
-    fun hideDatePicker()
-    fun showBottomSheet()
-    fun hideBottomSheet()
-    fun saveTask()
+    override fun updateTitle(title: String)
+    override fun updateDescription(description: String)
+    override fun updateSelectedDate(date: Long)
+    override fun updateSelectedPriority(priority: Priority)
+    override fun updateSelectedCategory(category: Category)
+    override fun showDatePicker()
+    override fun hideDatePicker()
+    override fun showBottomSheet()
+    override fun hideBottomSheet()
+    override fun saveTask()
     fun validateForm()
     fun clearMessages()
     fun loadTask(task: Task)
